requirements.txt
# Main
Django==3.0.4
django-environ==0.4.5
django-filter==2.2.0
django-debug-toolbar==2.2
mysqlclient==1.4.6

# Custom
btc-dev-tools==0.5.3
btc-template-tables==0.4.1
btc-action-set==0.2.3
btc-flex-forms==2.1
btc-floating-windows==1.0